==> lab2 code-1/Part_1.ipynb <==
import pandas as pd
import math

#Question 1:
#Import 3 datasets into Jupyter
calendar_df = pd.read_csv("calendar.csv")
listings_df = pd.read_csv("listings.csv")
reviews_df = pd.read_csv("reviews.csv")

#Question 2:
#Picked out the variables we want in the table and form a new dataframe
picked_var=listings_df[["host_response_rate", "host_acceptance_rate", "host_listings_count",
"host_total_listings_count", "accommodates", "bathrooms", "bedrooms", "beds", "price",
"weekly_price", "monthly_price", "security_deposit", "cleaning_fee", "guests_included",
"extra_people", "minimum_nights", "maximum_nights", "availability_30", "availability_90",
"availability_365", "number_of_reviews", "review_scores_rating", "review_scores_accuracy",
"review_scores_cleanliness", "review_scores_checkin", "review_scores_communication",
"review_scores_value", "reviews_per_month"]]




# Create 5 empty lists in order to build the table
max_l, min_l, mean_l, median_l, variance_l, std_l=[],[],[],[],[],[]

# Change the data type for variable with % sign
picked_var["host_response_rate"] = picked_var["host_response_rate"].str.rstrip('%').astype('float') 
picked_var["host_acceptance_rate"] = picked_var["host_acceptance_rate"].str.rstrip('%').astype('float') 


# Change the data type for variable with $ sign
def changet(cols):
    picked_var[cols]=picked_var[cols].dropna()
    picked_var[cols] = picked_var[cols].str.strip('$ ')
    picked_var[cols] = picked_var[cols].str.replace(',','').astype('float')
    return picked_var
    
changet("weekly_price")
changet("monthly_price")
changet("security_deposit")
changet("cleaning_fee")
changet("extra_people")
changet("price")

""" There are some values are badly measured. All measurements in review scores are measured in range from 0-10. 
However, for the host information like "host_response_rate" and "host_acceptance_rate", they are measure in percentage that range 
from 0%-100%."""
"""Another two measurements create confusions are "bedrooms" and "reviews per month". Bothe indicators are measured in float instead of 
integers. However, in the common sense, the numbers of bedrooms and reviews are in integer"""

"""In the dataset, some values include the measurements signs like % and $. In order to do the calculation, we take off these measurement signs."""

# Create a function that calculate the maximum, minimum, mean, median, variance and std. deviation for the variables in the new dataframe

def table(var):
    var=var.dropna()
   # var=np.log(var.astype('float'))
   # Max of the variable
    max_v=var.max()
    max_l.append(max_v)  # Put the calculated result in the empty list
   # Min of the variable
    min_v=var.min()
    min_l.append(min_v)
   # Mean of the variable
    mean = var.mean()
    mean_l.append(mean)
   # Median of the variable
    median= var.median()
    median_l.append(median)
   # Square deviations
    deviations = []
    for x in var:
        deviations.append((x-mean)**2)
    deviations = [(x - mean) ** 2 for x in var]
    # Variance
    variance = sum(deviations) / len(var)
    variance_l.append(variance)
    #Std Deviantion
    std_dev = math.sqrt(variance)
    std_l.append(std_dev)

# Calculate the values and put them in corresponding list.

# Create the index list
index_n=["host_response_rate", "host_acceptance_rate", "host_listings_count",
"host_total_listings_count", "accommodates", "bathrooms", "bedrooms", "beds", "price",
"weekly_price", "monthly_price", "security_deposit", "cleaning_fee", "guests_included",
"extra_people", "minimum_nights", "maximum_nights", "availability_30", "availability_90",
"availability_365", "number_of_reviews", "review_scores_rating", "review_scores_accuracy",
"review_scores_cleanliness", "review_scores_checkin", "review_scores_communication",
"review_scores_value", "reviews_per_month"]

#Calculate values for each variable
for i in index_n:
    table(picked_var[i])

# Build the table
#Create the column name list
col_n=["Variables","Minimum","Maximum","Mean","Median","Varaince","Std. Deviation"]

#Create the table by putting lists together
variable_df = pd.DataFrame(list(zip(index_n, min_l, max_l, mean_l, median_l,variance_l, std_l )), columns =col_n)
variable_df
